% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_exact.R
\name{pois.exact}
\alias{pois.exact}
\title{Confidence intervals for Poisson counts or rates}
\usage{
pois.exact(x, pt = 1, conf.level = 0.95)
}
\arguments{
\item{conf.level}{}
}
\value{

}
\description{
Confidence intervals for Poisson counts or rates
}
\examples{
pois.exact(x, pt = 1, conf.level = 0.95)
}
