% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_test.R
\name{rate_test}
\alias{rate_test}
\title{Test an event rate in two groups}
\usage{
rate_test(events_a, events_b, y_person_a, y_person_b)
}
\arguments{
\item{events_a}{number of events occurring the first group}

\item{events_b}{number of events occurring in the second group}

\item{y_person_a}{years (or months)-person for the first group}

\item{y_person_b}{years (or months)-person for the second group}
}
\value{
a tibble
}
\description{
Test an event rate in two groups
}
\examples{
rate_test(5,8,182.8,165.1)
}
