% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateratio_test.R
\name{rateratio.test}
\alias{rateratio.test}
\title{An Exact Rate Ratio Test Assuming Poisson Counts
Performs the uniformy most powerful unbiased test on the ratio of rates of two Poisson counts with given time (e.g., perons-years) at risk for each count.}
\usage{
rateratio.test(
  x,
  n,
  RR = 1,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{conf.level}{}
}
\value{
An object of class ‘htest’
}
\description{
An Exact Rate Ratio Test Assuming Poisson Counts
Performs the uniformy most powerful unbiased test on the ratio of rates of two Poisson counts with given time (e.g., perons-years) at risk for each count.
}
\examples{
rateratio.test(c(2,9),c(17877,16660))

}
